Replace WelcomePage with LoginPage, remove unused react-toastify dependency, and update user-related types and services
