+++ src/components/layouts/users/ActionsUsers.tsx
@@ -1,61 +1,3 @@
-// "use client";
-
-// import { Row } from "@tanstack/react-table";
-// import {
-//   Sheet,
-//   SheetContent,
-//   SheetDescription,
-//   SheetHeader,
-//   SheetTitle,
-//   SheetTrigger,
-// } from "@/components/ui/sheet";
-// import { Button } from "@/components/ui/button";
-// import { IconTrash } from "@tabler/icons-react";
-// import { useDeleteUser } from "@/hooks/Users/useUserMutations";
-// import { toast } from "@/hooks/use-toast"; // Tu toast personalizado
-
-// interface ProjectActionsProps<TData> {
-//   row: Row<TData>;
-// }
-
-// export function ActionUser<TData>({ row }: ProjectActionsProps<TData>) {
-//   const { mutateAsync } = useDeleteUser();
-
-//   const handleDelete = async () => {
-//     if (window.confirm("¿Estás seguro de eliminar este usuario?")) {
-//       try {
-//         await mutateAsync(row.getValue("id_user"));
-//         toast({ title: "Éxito", description: "Usuario eliminado correctamente", open: true });
-//       } catch (error: any) {
-//         toast({ title: "Error", description: error.message, open: true });
-//       }
-//     }
-//   };
-
-//   return (
-//     <Sheet>
-//       <div className="flex items-center justify-center">
-//         <SheetTrigger>
-//           <IconTrash className="cursor-pointer text-red-500" />
-//         </SheetTrigger>
-//       </div>
-//       <SheetContent className="w-[200px] sm:w-[540px]">
-//         <SheetHeader>
-//           <SheetTitle>Usuario: {row.getValue("name")}</SheetTitle>
-//           <SheetDescription>
-//             Esta acción eliminará permanentemente al usuario.
-//           </SheetDescription>
-//         </SheetHeader>
-//         <div className="flex justify-end mt-4">
-//           <Button variant="destructive" onClick={handleDelete} className="bg-red-600 text-white hover:bg-red-700">
-//             Eliminar
-//           </Button>
-//         </div>
-//       </SheetContent>
-//     </Sheet>
-//   );
-// }
-
 "use client";
 
 import { Row } from "@tanstack/react-table";
